Move event processor, factory and related traits to service module
