+++ nexus-watcher/src/service/traits/mod.rs
