+++ nexus-watcher/tests/event_processor/tags/fail_index.rs
@@ -1,7 +1,7 @@
 use crate::event_processor::utils::watcher::{retrieve_and_handle_event_line, WatcherTest};
 use anyhow::{anyhow, Result};
 use chrono::Utc;
-use nexus_watcher::events::TEventProcessorFactory;
+use nexus_watcher::service::TEventProcessorFactory;
 use pubky::Keypair;
 use pubky_app_specs::{traits::HashId, PubkyAppPost, PubkyAppTag, PubkyAppUser};
 use tracing::error;
