+++ nexus-watcher/tests/event_processor/utils/watcher.rs
@@ -7,10 +7,10 @@ use nexus_common::models::homeserver::Homeserver;
 use nexus_common::types::DynError;
 use nexus_watcher::events::retry::event::RetryEvent;
 use nexus_watcher::events::Event;
-use nexus_watcher::events::EventProcessorFactory;
 use nexus_watcher::events::Moderation;
-use nexus_watcher::events::TEventProcessorFactory;
+use nexus_watcher::service::EventProcessorFactory;
 use nexus_watcher::service::NexusWatcher;
+use nexus_watcher::service::TEventProcessorFactory;
 use pubky::Keypair;
 use pubky::PublicKey;
 use pubky_app_specs::PubkyId;
